Moved consensus and fsm to cluster folder
