+++ cluster/consensus/fsm/delete.go
