+++ cluster/consensus/logger.go
